+++ Cryptosystem/src/test/CryptosystemTest.java
@@ -1,7 +1,5 @@
 package test;
 
-import java.math.BigInteger;
-
 import decrypt.CiphertextDecrypter;
 import encrypt.PlaintextEncrypter;
 import keyGen.KeyTuple;
@@ -25,20 +23,20 @@ public class CryptosystemTest {
 	 * tests all of the functionality of a typical success scenario for the entire system. Informal. Missing tests include tests for exceeding the parameters given in the paper.
 	 */
 	private static void overallTest(){
-		BigInteger testSecurityParameter = BigInteger.valueOf(5000);
-		BigInteger testModulus = BigInteger.valueOf(7);
+		double testSecurityParameter = 5000;
+		double testModulus = 7;
 		
 		//needs to me modified to fit trapsamp...
 		KeyTuple outputTuple = TrapdoorSampler.trapSamp(testSecurityParameter, testModulus);
 
-		BigInteger[][] privateKey = outputTuple.getPrivateKey();
-		BigInteger[][] publicKey = outputTuple.getPublicKey();
+		double[][] privateKey = outputTuple.getPrivateKey();
+		double[][] publicKey = outputTuple.getPublicKey();
 		
 		//Initialise plaintext matrix. Should make random eventually, so that we don't keep using the same test values
-		BigInteger[][] plainText = new BigInteger[10][10];
+		double[][] plainText = new double[10][10];
 		for (int x=0; x<10;x++){
 			for (int y=0;y<10;y++){
-				plainText[x][y] = BigInteger.valueOf(x+y);
+				plainText[x][y] = x+y;
 			}			
 		}
 		
@@ -47,7 +45,7 @@ public class CryptosystemTest {
 		System.out.println();
 		
 		PlaintextEncrypter pe = new PlaintextEncrypter();
-		BigInteger[][] cipherText = pe.encrypt(plainText, publicKey);
+		double[][] cipherText = pe.encrypt(plainText, publicKey);
 		
 		System.out.println("Ciphertext:");
 		System.out.println(cipherText);
@@ -55,13 +53,13 @@ public class CryptosystemTest {
 		
 		HomomorphicOps ho = new HomomorphicOps();
 		
-		BigInteger[][] addedCipherText = ho.add(cipherText, cipherText);
+		double[][] addedCipherText = ho.add(cipherText, cipherText);
 		
 		System.out.println("Added Ciphertext:");
 		System.out.println(addedCipherText);
 		System.out.println();
 		
-		BigInteger[][] multipliedCipherText = ho.multiply(cipherText, cipherText);
+		double[][] multipliedCipherText = ho.multiply(cipherText, cipherText);
 		
 		System.out.println("Multiplied Ciphertext:");
 		System.out.println(multipliedCipherText);
@@ -69,9 +67,9 @@ public class CryptosystemTest {
 		
 		CiphertextDecrypter cd = new CiphertextDecrypter();
 		
-		BigInteger[][] decryptedCiphertext = cd.decrypt(cipherText, privateKey);
-		BigInteger[][] decryptedAddedCiphertext = cd.decrypt(addedCipherText, privateKey);
-		BigInteger[][] decryptedMultipliedCiphertext = cd.decrypt(multipliedCipherText, privateKey);
+		double[][] decryptedCiphertext = cd.decrypt(cipherText, privateKey);
+		double[][] decryptedAddedCiphertext = cd.decrypt(addedCipherText, privateKey);
+		double[][] decryptedMultipliedCiphertext = cd.decrypt(multipliedCipherText, privateKey);
 		
 		System.out.println("Decrypted Plaintext:");
 		System.out.println(decryptedCiphertext);
